NRFX-7881: fix waiting for STOPPED event when it does not arrive
Add fix for waiting for SPIM STOPPED event when it does not arrive.
When triggering task STOP in SPIM when no transaction is going,
event STOPPED does not arrive, even though it should according
to Product Specification. Add checking if any transfer is in
progress before task STOP is triggered.

Signed-off-by: Magdalena Pastula <magdalena.pastula@nordicsemi.no>